+++ CHANGELOG.rst
@@ -3,6 +3,7 @@
   - Fix mypy errors: vivarium/framework/lifecycle.py
   - Fix mypy errors: vivarium/framework/population/manager.py
   - Fix mypy errors: vivarium/framework/population/population_view.py
+  - Enable population manager and population view methods to accept strings
 
 **3.0.6 - 09/20/24**