+++ CHANGELOG.rst
@@ -4,6 +4,7 @@
   - Fix mypy errors: vivarium/framework/population/manager.py
   - Fix mypy errors: vivarium/framework/population/population_view.py
   - Enable population manager and population view methods to accept strings
+  - Fix mypy errors: vivarium/framework/plugins.py
 
 **3.0.6 - 09/20/24**